+++ erp-model/erp-model-plm/src/main/java/com/erp/model/plm/dto/ProductSaleShowDTO.java
@@ -41,7 +41,7 @@ public class ProductSaleShowDTO implements Serializable {
     /**
      * 年目标销售量
      */
-    private Integer yearSaleQty;
+    private Long yearSaleQty;
 
     /**
      * 年目标销售额
